+++ examples/traitless.rs
@@ -15,19 +15,13 @@ fn main() {
 }
 
 fn setup(mut commands: Commands, asset_server: Res<AssetServer>) {
-    commands.spawn(Camera2dBundle::default());
+    commands.spawn(Camera2d);
 
     commands.spawn(LdtkWorldBundle {
-        ldtk_handle: asset_server.load("my_project.ldtk"),
+        ldtk_handle: asset_server.load("my_project.ldtk").into(),
         ..Default::default()
     });
 }
-
-#[derive(Default, Component)]
-struct ComponentA;
-
-#[derive(Default, Component)]
-struct ComponentB;
 
 fn process_my_entity(
     mut commands: Commands,
@@ -41,7 +35,7 @@ fn process_my_entity(
 
             if let Some(tile) = &entity_instance.tile {
                 let layout = texture_atlases.add(TextureAtlasLayout::from_grid(
-                    Vec2::new(tile.w as f32, tile.h as f32),
+                    UVec2::new(tile.w as u32, tile.h as u32),
                     16,
                     95,
                     None,
@@ -53,12 +47,9 @@ fn process_my_entity(
                     layout,
                 };
 
-                commands.entity(entity).insert(SpriteSheetBundle {
-                    atlas,
-                    texture,
-                    transform: *transform,
-                    ..Default::default()
-                });
+                commands
+                    .entity(entity)
+                    .insert((Sprite::from_atlas_image(texture, atlas), *transform));
             }
         }
     }