+++ macros/src/lib.rs
@@ -6,13 +6,14 @@ mod ldtk_int_cell;
 #[proc_macro_derive(
     LdtkEntity,
     attributes(
-        sprite_bundle,
-        sprite_sheet_bundle,
+        sprite,
+        sprite_sheet,
         worldly,
         grid_coords,
         ldtk_entity,
         from_entity_instance,
-        with
+        with,
+        default,
     )
 )]
 pub fn ldtk_entity_derive(input: TokenStream) -> TokenStream {
@@ -21,7 +22,10 @@ pub fn ldtk_entity_derive(input: TokenStream) -> TokenStream {
     ldtk_entity::expand_ldtk_entity_derive(ast)
 }
 
-#[proc_macro_derive(LdtkIntCell, attributes(ldtk_int_cell, from_int_grid_cell, with))]
+#[proc_macro_derive(
+    LdtkIntCell,
+    attributes(ldtk_int_cell, from_int_grid_cell, with, default)
+)]
 pub fn ldtk_int_cell_derive(input: TokenStream) -> TokenStream {
     let ast = syn::parse(input).unwrap();