+++ src/assets/ldtk_project.rs
@@ -7,10 +7,9 @@ use crate::{
     ldtk::{raw_level_accessor::RawLevelAccessor, LdtkJson, Level},
 };
 use bevy::{
-    asset::{io::Reader, AssetLoader, AssetPath, AsyncReadExt, LoadContext},
+    asset::{io::Reader, AssetLoader, AssetPath, LoadContext},
     prelude::*,
     reflect::Reflect,
-    utils::BoxedFuture,
 };
 use derive_getters::Getters;
 use derive_more::From;
@@ -240,95 +239,92 @@ impl AssetLoader for LdtkProjectLoader {
     type Settings = LdtkProjectLoaderSettings;
     type Error = LdtkProjectLoaderError;
 
-    fn load<'a>(
-        &'a self,
-        reader: &'a mut Reader,
-        settings: &'a Self::Settings,
-        load_context: &'a mut LoadContext,
-    ) -> BoxedFuture<'a, Result<Self::Asset, Self::Error>> {
-        Box::pin(async move {
-            // Create custom loader with my algorithm in a custom laoder
-            let mut bytes = Vec::new();
-            reader.read_to_end(&mut bytes).await?;
-            let mut data: LdtkJson = serde_json::from_slice(&bytes)?;
+    async fn load(
+        &self,
+        reader: &mut dyn Reader,
+        settings: &Self::Settings,
+        load_context: &mut LoadContext<'_>,
+    ) -> Result<Self::Asset, Self::Error> {
+        let mut bytes = Vec::new();
+        reader.read_to_end(&mut bytes).await?;
+        let mut data: LdtkJson = serde_json::from_slice(&bytes)?;
 
-            let transform_data = self
+        let transform_data = self
                 .callback
                 .as_ref()
                 .map(|callback| callback(data.clone(), settings.data.clone()));
-            if transform_data.is_some() {
-                data = transform_data.unwrap();
-            }
+        if transform_data.is_some() {
+            data = transform_data.unwrap();
+        }
 
-            let mut tileset_map: HashMap<i32, Handle<Image>> = HashMap::new();
-            for tileset in &data.defs.tilesets {
-                if let Some(tileset_path) = &tileset.rel_path {
-                    let asset_path = ldtk_path_to_asset_path(load_context.path(), tileset_path);
+        let mut tileset_map: HashMap<i32, Handle<Image>> = HashMap::new();
+        for tileset in &data.defs.tilesets {
+            if let Some(tileset_path) = &tileset.rel_path {
+                let asset_path = ldtk_path_to_asset_path(load_context.path(), tileset_path);
 
-                    tileset_map.insert(tileset.uid, load_context.load(asset_path));
-                } else if tileset.embed_atlas.is_some() {
-                    warn!("Ignoring LDtk's Internal_Icons. They cannot be displayed due to their license.");
-                } else {
-                    let identifier = &tileset.identifier;
-                    warn!("{identifier} tileset cannot be loaded, it has a null relative path.");
-                }
-            }
-
-            let int_grid_image_handle = data
-                .defs
-                .create_int_grid_image()
-                .map(|image| load_context.add_labeled_asset("int_grid_image".to_string(), image));
-
-            let ldtk_project = if data.external_levels {
-                #[cfg(feature = "external_levels")]
-                {
-                    let mut level_map = HashMap::new();
-
-                    for (level_indices, level) in data.iter_raw_levels_with_indices() {
-                        let level_metadata =
-                            load_external_level_metadata(load_context, level_indices, level)?;
-
-                        level_map.insert(level.iid.clone(), level_metadata);
-                    }
-
-                    LdtkProject::new(
-                        LdtkProjectData::Parent(LdtkJsonWithMetadata::new(data, level_map)),
-                        tileset_map,
-                        int_grid_image_handle,
-                    )
-                }
-
-                #[cfg(not(feature = "external_levels"))]
-                {
-                    Err(LdtkProjectLoaderError::ExternalLevelsDisabled)?
-                }
+                tileset_map.insert(tileset.uid, load_context.load(asset_path));
+            } else if tileset.embed_atlas.is_some() {
+                warn!("Ignoring LDtk's Internal_Icons. They cannot be displayed due to their license.");
             } else {
-                #[cfg(feature = "internal_levels")]
-                {
-                    let mut level_map = HashMap::new();
+                let identifier = &tileset.identifier;
+                warn!("{identifier} tileset cannot be loaded, it has a null relative path.");
+            }
+        }
 
-                    for (level_indices, level) in data.iter_raw_levels_with_indices() {
-                        let level_metadata =
-                            load_level_metadata(load_context, level_indices, level, true)?;
+        let int_grid_image_handle = data
+            .defs
+            .create_int_grid_image()
+            .map(|image| load_context.add_labeled_asset("int_grid_image".to_string(), image));
 
-                        level_map.insert(level.iid.clone(), level_metadata);
-                    }
+        let ldtk_project = if data.external_levels {
+            #[cfg(feature = "external_levels")]
+            {
+                let mut level_map = HashMap::new();
 
-                    LdtkProject::new(
-                        LdtkProjectData::Standalone(LdtkJsonWithMetadata::new(data, level_map)),
-                        tileset_map,
-                        int_grid_image_handle,
-                    )
+                for (level_indices, level) in data.iter_raw_levels_with_indices() {
+                    let level_metadata =
+                        load_external_level_metadata(load_context, level_indices, level)?;
+
+                    level_map.insert(level.iid.clone(), level_metadata);
                 }
 
-                #[cfg(not(feature = "internal_levels"))]
-                {
-                    Err(LdtkProjectLoaderError::InternalLevelsDisabled)?
-                }
-            };
+                LdtkProject::new(
+                    LdtkProjectData::Parent(LdtkJsonWithMetadata::new(data, level_map)),
+                    tileset_map,
+                    int_grid_image_handle,
+                )
+            }
 
-            Ok(ldtk_project)
-        })
+            #[cfg(not(feature = "external_levels"))]
+            {
+                Err(LdtkProjectLoaderError::ExternalLevelsDisabled)?
+            }
+        } else {
+            #[cfg(feature = "internal_levels")]
+            {
+                let mut level_map = HashMap::new();
+
+                for (level_indices, level) in data.iter_raw_levels_with_indices() {
+                    let level_metadata =
+                        load_level_metadata(load_context, level_indices, level, true)?;
+
+                    level_map.insert(level.iid.clone(), level_metadata);
+                }
+
+                LdtkProject::new(
+                    LdtkProjectData::Standalone(LdtkJsonWithMetadata::new(data, level_map)),
+                    tileset_map,
+                    int_grid_image_handle,
+                )
+            }
+
+            #[cfg(not(feature = "internal_levels"))]
+            {
+                Err(LdtkProjectLoaderError::InternalLevelsDisabled)?
+            }
+        };
+
+        Ok(ldtk_project)
     }
 
     fn extensions(&self) -> &[&str] {
@@ -340,7 +336,7 @@ impl AssetLoader for LdtkProjectLoader {
 mod tests {
     use super::*;
     use derive_more::Constructor;
-    use fake::{Dummy, Fake, Faker};
+    use fake::{uuid::UUIDv4, Dummy, Fake};
     use rand::Rng;
 
     #[derive(Copy, Clone, Debug, Default, PartialEq, Eq, Constructor)]
@@ -362,13 +358,22 @@ mod tests {
                 .defs
                 .tilesets
                 .iter()
-                .map(|tileset| (tileset.uid, Handle::weak_from_u128(Faker.fake())))
+                .map(|tileset| {
+                    (
+                        tileset.uid,
+                        Handle::Weak(AssetId::Uuid {
+                            uuid: UUIDv4.fake(),
+                        }),
+                    )
+                })
                 .collect();
 
             LdtkProject {
                 data,
                 tileset_map,
-                int_grid_image_handle: Some(Handle::weak_from_u128(Faker.fake())),
+                int_grid_image_handle: Some(Handle::Weak(AssetId::Uuid {
+                    uuid: UUIDv4.fake(),
+                })),
             }
         }
     }