+++ src/utils.rs
@@ -254,7 +254,9 @@ pub(crate) fn set_all_tiles_with_func(
                 .map(|tile_bundle| commands.spawn(tile_bundle).insert(tilemap_id).id());
             match tile_entity {
                 Some(tile_entity) => storage.set(&tile_pos, tile_entity),
-                None => storage.remove(&tile_pos),
+                None => {
+                    storage.remove(&tile_pos);
+                }
             }
         }
     }
@@ -305,27 +307,27 @@ where
     try_each_optional_permutation(a, b, |x, y| map.get(&(x, y))).unwrap_or(default)
 }
 
-/// Creates a [SpriteSheetBundle] from the entity information available to the
+/// Creates a [`Sprite`] with [`TextureAtlas`] from the entity information available to the
 /// [LdtkEntity::bundle_entity] method.
 ///
-/// Used for the `#[sprite_sheet_bundle]` attribute macro for `#[derive(LdtkEntity)]`.
-/// See [LdtkEntity#sprite_sheet_bundle] for more info.
-pub fn sprite_sheet_bundle_from_entity_info(
+/// Used for the `#[sprite_sheet]` attribute macro for `#[derive(LdtkEntity)]`.
+/// See [LdtkEntity#sprite_sheet] for more info.
+pub fn sprite_sheet_from_entity_info(
     entity_instance: &EntityInstance,
     tileset: Option<&Handle<Image>>,
     tileset_definition: Option<&TilesetDefinition>,
     texture_atlases: &mut Assets<TextureAtlasLayout>,
     grid: bool,
-) -> SpriteSheetBundle {
+) -> Sprite {
     match (tileset, &entity_instance.tile, tileset_definition) {
-        (Some(tileset), Some(tile), Some(tileset_definition)) => SpriteSheetBundle {
-            atlas: if grid {
+        (Some(tileset), Some(tile), Some(tileset_definition)) => {
+            let texture_atlas = if grid {
                 let layout = TextureAtlasLayout::from_grid(
-                    Vec2::new(tile.w as f32, tile.h as f32),
-                    tileset_definition.c_wid as usize,
-                    tileset_definition.c_hei as usize,
-                    Some(Vec2::splat(tileset_definition.spacing as f32)),
-                    Some(Vec2::splat(tileset_definition.padding as f32)),
+                    UVec2::new(tile.w as u32, tile.h as u32),
+                    tileset_definition.c_wid as u32,
+                    tileset_definition.c_hei as u32,
+                    Some(UVec2::splat(tileset_definition.spacing as u32)),
+                    Some(UVec2::splat(tileset_definition.padding as u32)),
                 );
                 let texture_atlas: Handle<TextureAtlasLayout> = texture_atlases.add(layout);
                 TextureAtlas {
@@ -335,50 +337,47 @@ pub fn sprite_sheet_bundle_from_entity_info(
                         + (tile.x / (tile.w + tileset_definition.spacing)) as usize,
                 }
             } else {
-                let mut layout = TextureAtlasLayout::new_empty(Vec2::new(
-                    tileset_definition.px_wid as f32,
-                    tileset_definition.px_hei as f32,
+                let mut layout = TextureAtlasLayout::new_empty(UVec2::new(
+                    tileset_definition.px_wid as u32,
+                    tileset_definition.px_hei as u32,
                 ));
-                layout.add_texture(Rect::new(
-                    tile.x as f32,
-                    tile.y as f32,
-                    (tile.x + tile.w) as f32,
-                    (tile.y + tile.h) as f32,
+                layout.add_texture(URect::new(
+                    tile.x as u32,
+                    tile.y as u32,
+                    (tile.x + tile.w) as u32,
+                    (tile.y + tile.h) as u32,
                 ));
                 let texture_atlas: Handle<TextureAtlasLayout> = texture_atlases.add(layout);
                 TextureAtlas {
                     layout: texture_atlas,
                     index: 0,
                 }
-            },
-            texture: tileset.clone(),
-            ..Default::default()
-        },
+            };
+
+            Sprite::from_atlas_image(tileset.clone(), texture_atlas)
+        }
         _ => {
-            warn!("EntityInstance needs a tile, an associated tileset, and an associated tileset definition to be bundled as a SpriteSheetBundle");
-            SpriteSheetBundle::default()
+            warn!("EntityInstance needs a tile, an associated tileset, and an associated tileset definition to be inserted as a Sprite");
+            Sprite::default()
         }
     }
 }
 
-/// Creates a [SpriteBundle] from the entity information available to the
+/// Creates a [Sprite] from the entity information available to the
 /// [LdtkEntity::bundle_entity] method.
 ///
-/// Used for the `#[sprite_bundle]` attribute macro for `#[derive(LdtkEntity)]`.
-/// See [LdtkEntity#sprite_bundle] for more info.
-pub fn sprite_bundle_from_entity_info(tileset: Option<&Handle<Image>>) -> SpriteBundle {
+/// Used for the `#[sprite]` attribute macro for `#[derive(LdtkEntity)]`.
+/// See [LdtkEntity#sprite] for more info.
+pub fn sprite_from_entity_info(tileset: Option<&Handle<Image>>) -> Sprite {
     let tileset = match tileset {
         Some(tileset) => tileset.clone(),
         None => {
-            warn!("EntityInstance needs a tileset to be bundled as a SpriteBundle");
-            return SpriteBundle::default();
+            warn!("EntityInstance needs a tileset to be inserted as a Sprite");
+            return Sprite::default();
         }
     };
 
-    SpriteBundle {
-        texture: tileset,
-        ..Default::default()
-    }
+    Sprite::from_image(tileset)
 }
 
 #[cfg(test)]